+++ tests/commands/test_test.py
@@ -110,6 +110,60 @@ def test_package_not_exist():
         assert '`{}` is not an editable package.'.format('ok') in result.output
 
 
+def test_local():
+    with temp_chdir() as d:
+        runner = CliRunner()
+        runner.invoke(hatch, ['new', 'ok', '--basic'])
+        package_dir = os.path.join(d, 'ok')
+        create_test_passing(package_dir)
+
+        venv_dir = os.path.join(d, 'venv')
+        create_venv(venv_dir)
+
+        with venv(venv_dir):
+            install_packages(['-e', package_dir])
+            result = runner.invoke(hatch, ['test', '-l', '-g'])
+
+        assert result.exit_code == 0
+        assert 'Package `ok` has been selected.' in result.output
+        assert '1 passed' in result.output
+
+
+def test_local_not_exist():
+    with temp_chdir() as d:
+        runner = CliRunner()
+        venv_dir = os.path.join(d, 'venv')
+        create_venv(venv_dir)
+
+        with venv(venv_dir):
+            result = runner.invoke(hatch, ['test', '-l', '-g'])
+
+        assert result.exit_code == 1
+        assert 'There are no local packages available.' in result.output
+
+
+def test_local_multiple():
+    with temp_chdir() as d:
+        runner = CliRunner()
+        runner.invoke(hatch, ['new', 'ok', '--basic'])
+        runner.invoke(hatch, ['new', 'ko', '--basic'])
+
+        venv_dir = os.path.join(d, 'venv')
+        create_venv(venv_dir)
+
+        with venv(venv_dir):
+            install_packages(['-e', os.path.join(d, 'ok')])
+            install_packages(['-e', os.path.join(d, 'ko')])
+
+            result = runner.invoke(hatch, ['test', '-l', '-g'])
+
+        assert result.exit_code == 1
+        assert (
+            'There are multiple local packages available. '
+            'Select one with the optional argument.'
+        ) in result.output
+
+
 def test_path_relative():
     with temp_chdir() as d:
         runner = CliRunner()